+++ src/BecklynRadBundle.php
@@ -1,9 +1,9 @@
 <?php declare(strict_types=1);
 
-namespace Becklyn\RadBundle;
+namespace Becklyn\Rad;
 
-use Becklyn\RadBundle\Bundle\BundleExtension;
-use Becklyn\RadBundle\DependencyInjection\DoctrineExtensionsCompilerPass;
+use Becklyn\Rad\DependencyInjection\DoctrineExtensionsCompilerPass;
+use Becklyn\RadBundles\Bundle\BundleExtension;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;