+++ src/Controller/DownloadableStringTrait.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Becklyn\RadBundle\Controller;
+namespace Becklyn\Rad\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\ControllerTrait;
 use Symfony\Component\HttpFoundation\HeaderUtils;