+++ src/Search/SimpleEntitySearchHelper.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Becklyn\RadBundle\Search;
+namespace Becklyn\Rad\Search;
 
 use Doctrine\ORM\Query\Expr\Comparison;
 use Doctrine\ORM\Query\Expr\Func;