+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 *   (bc) Removed `InvalidSortOperationException`.
 *   (bc) Removed `BundleExtension` and `ConfigurableBundleExtension`.
 *   (internal) Renamed Composer package from `becklyn/rad-bundle` to `becklyn/rad`. See README for more information.
+*   (bc) Rename namespace `Becklyn\RadBundle` to `Becklyn\Rad`.
 
 
 7.13.2